Change where the cdnjs cache is saved
This should fix the issues in #24 and #25

Hoo boy this was a pig of an issue:

For a while I was unable to reproduce this issue since it worked fine for me in Sublime Text 3 on Windows 8. As soon as I tried to use it in Sublime Text 2 however, the cache was trying to save itself in system32! I have no idea how ST2 thought the package was there and I tried many things to get it back to where the package is. Unfortunately I had no such luck

So I've moved where the cache is being saved completely. It should be in the home folder of whatever operating system you're using (in Windows this would be something like C:\Users\YourName\)